+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
+- `flipperzero_sys::furi::UnsafeRecord::open` now takes a `&'static CStr`
+
 ### Removed
 
 ## [0.13.0]